Update boedeltermen.csv
Reviewed data. Added (provisional) URIs to curated terms. Added 2 new categories: 'vee' and 'rest'. The 'rest' category is for terms that either do not fit in the current scheme or need further scrutiny. The rest category is also used for corrections of some HTR errors. Terms under 'rest' are not URIfied.